Refactor settings into separate modules
