+++ indico/core/db/sqlalchemy/principals.py
